Add a PathTemplate type to create templates paths
If a template contains keys that can be empty, the rendered string can
alter the type of path created. For instance, given the template

    "{subdirectory}/demo_file_name"

If subdirectory is replaced with an empty string, the path created from
the rendered string becomes absolute whereas if it has any other value,
it is relative. This causes issues when joining paths from different
templates.

To protect against this, the template is split into path segments at
creation and each segment is rendered individually. If a segment is
empty, it does not modify the path.