Add Detector type to ensure detector names are path-safe
If a detector named '../other/' is passed to a template containing
"{detector}" as a standalone path segment, it would result in a path
outside the intended directory. For now detector names are normalized by
replacing any '/' or '.' with '_'. This may not be required in every
case but lets it stay consistent.