Rewrite SqliteScanPathService
Motivated by several problems with the previous implementation.

There were issues with the previous queries around getting the all the
latest templates for a beamline in a single query (where non-latest
results would sometimes be returned). As graphql picks and chooses which
templates are required at runtime anyway, it makes sense to return each
template individually.

The queries are now extracted into text files in the `queries`
directory. These are still built into the executable and the types are
still checked at compile time but it keeps the rust source cleaner and
the SQL all in one place.

The new template methods each return a PathTemplate instead of the raw
string. This means that callers don't need to build their own templates
or deal with the errors from each stage (unless they need to - the
underlying errors are still available).