Make numtracker functions actually async
The trait next_scan_number method was marked as async even though
nothing it called required it. The methods it calls are now async using
tokio::fs and the imports for std::fs items have been replaced with
qualified paths so it is clear which is being used for each call.