Add sync subcommand
Includes (optional) subcommands to either import or export the scan
numbers from the DB and the configured directories. If neither is
specified, the user is prompted for which side to update.

By default, the user is prompted before any change but there is a
'--force' flag to run everything automatically if either import or
export is specified.