Raise error if no scan number is incremented
If there is no entry in the scan number table for a given beamline, the
'increment scan number' query becomes a no-op but doesn't fail. Checking
the number of rows updated allows the error to be propagated.

The inverse (wheree there are too many entries for a beamline should not
be possible if constraints are correct but handle it anyway to avoid
runtime panics.