Reduce schema to a single table
While having additional tables for templates allowed multiple beamlines
to share a template, in practice this was not likely to be useful and
would not allow individual beamlines to easily update.

The mapping from beamline to template in use allowed the history of
templates to be stored but not in a way that would have been useful, and
definitely not useful enough to justify the additional complexity.

In future it would be good to strengthen the constraints on the
remaining table to ensure that templates are always set (as it doesn't
make sense for a beamline to be present with no templates) but this can
be added in a later change - this keeps to the previous behaviour
where possible.