Almost complete rewrite to simplify everything
Replace 'query per field' approach to accessing data with a single query
to return the full configuration for a beamline. In most cases, reduces
the number of queries per graphQL request and greatly simplifies the
db_service module.

Replace the `info` CLI command with a `configuration` graphQL query.

Replace the `configuration` CLI command with a `configure` graphQL
mutation.

Replace the `sync` CLI command with a redesigned integration with
directory number tracking. The DB and directory will be left in sync
after every request for the next scan number with the highest of the two
values being taken as the latest value.

File locking has been removed for the directory based number tracking as
it did not serve any real benefit. External processes would not be
respecting the lock and race conditions would still be present.