Return status information from status endpoint
The status endpoint is currently only used as a liveness probe for
kubernetes deployments but returning a version and start time provides a
useful sanity check of the running version without having to use the
kubernetes dashboard.

The dockerfile needs to be updated to include the build.rs file and set
the workdir. Previously, cargo was running in the root directory and
would complain about infinite file system loops in /proc/.