Bundle root certificates at compile time
In order to make https connections for authorisation, the reqwuest
library needs root certificates. In the minimal container used, these
are not present in the OS so authorisation fails. By bundling them at
compile time (via the webpki-roots crate) the native certificates are
not required.

This usually presents a maintenance burden as the application needs to
be recompiled if root CA certificates change, however in the case where
the application is running in a container, the container would need to
be rebuilt in either case to include the new certificates.