Wrap auth request input as field in JSON query
The input should be sent to the OPA instance as the 'input' field
of the query data instead of as the data itself.

The tests have also been updated to reflect this and now use the
expected json when mocking the auth server instead of relying on the
same request/response types as the implementation so that
inconsistencies are easier to catch in future.