Use a temp file for read-only DB tests
An in-memory DB does not allow a mix of read-only and read-write
connections. The initial connection has to have write access to set up
the initial schema so subsequent connections all end up having write
access as well, even when setting read_only(true) in the pool options.

The read_only test previously passed as the connection in the test was
seeing a new in-memory database and the test didn't distinguish between
'DB is read-only' and 'Table does not exist' error cases.