Check for changed files during the CI tests
If dependencies are updated by a change, the lock file should be updated
to match. If it is out of sync it will be updated during the build and
be left as an unstaged change. If there are any unstaged changes at the
end of the build, the CI should now fail.