Client subcommand for interacting with service (#104)
Provides a more convenient way of generating and passing auth tokens and
saves writing graphql queries in the terminal to test things using curl.

New features are behind the `client` subcommand which itself has three
further commands:
* configuration: Show/query existing configurations
* configure: Update/add new configurations
* visit-directory: Generate the visit directory for the given instrument

As part of the changes, the '--db' option has been made specific to the
serve command as it is not required elsewhere.