Split format and clippy into separate CI jobs
Improve CI performance by making cache more effective.

The format check requires a nightly toolchain to support unstable
formatting options but does not need to compile anything - this means
the cache doesn't really affect the runtime and can be skipped.

The clippy job needs to compile the project and therefore needs all the
dependencies available.

The cache uses the version of rust as part of the key so a nightly build
guarantees a cache miss for the first job each day. Using the stable
toolchain for the clippy check allows it to use the same cache from
previous runs reducing the need to download crates twice for each PR.