Add signal handler and shutdown hook
When the application is running in a container, it is given PID 1 which
means it does not automatically get given signal handlers and can't be
killed. When running the container in the foreground this stops Ctrl-C
from interrupting it while in kubernetes it means the graceful shutdown
always times out resulting in the container eventually being forcibly
deleted, killing the process.

Adding an explicit handler for the common signals used to interrupt
processes means that Ctrl-C, kubernetes and `podman stop` all work as
expected.